pyproject.toml
[project]
name = "remove-bg"
version = "0.1.0"
description = "Automatically remove image backgrounds using Python"
authors = [
  { name = "andersonlimacrv", email = "andersonlimacrv@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
  "rembg>=2.0.69,<3.0.0",
  "pillow>=12.0.0,<13.0.0",
  "onnxruntime (>=1.23.2,<2.0.0)"
]

[project.scripts]
rmbg = "app.main:main"

[tool.poetry]
packages = [{ include = "app" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
